fix: extract renderClientShell to routes/html.ts, fix e2e test failures
- Extract renderClientShell from program.ts into host/src/routes/html.ts
  (nonce now passed as param instead of recomputing CSP_STRICT)
- Fix stale theme toggle locator: 'Toggle theme' -> 'Switch to ... theme'
- Fix proxy target json-proxy-target.ts:
  - Return oRPC-compatible RPC format { json, meta } for /api/rpc/*
  - Add stubs for /api/auth/get-session and /api/auth/near/list-accounts
- Fix 18 host vitest failures: provide mock session in test render options
- Replace stale assertion in ssr-bundled-runtime test
- Document bun run --cwd host test in AGENTS.md
